refactor: Removing circular dependency from SiteModeration state
Refs: #10
