+++ web-application/client/state/lib/relations.js
@@ -8,7 +8,7 @@ import { setPostsInDictionary, removePost } from '/state/Post/slice'
 import { setPostCommentsInDictionary, removePostComment } from '/state/PostComment/slice'
 import { setPostReactionsInDictionary, removePostReaction } from '/state/PostReaction/slice'
 import { setPostSubscriptionsInDictionary, removePostSubscription } from '/state/PostSubscription/slice'
-import { setSiteModerationsInDictionary, removeSiteModeration } from '/state/admin/siteModeration'
+import { setSiteModerationsInDictionary, removeSiteModeration } from '/state/SiteModeration/slice'
 import { setUsersInDictionary, removeUser } from '/state/User/slice'
 import { setUserRelationshipsInDictionary, removeUserRelationship } from '/state/UserRelationship/slice'
